some typos

+++ APA102.h
@@ -9,7 +9,7 @@ typedef struct PixelInfo {
 
 class APA102 {
     public:
-        APA102(gpio_num_t mosiPin, gpio_num_t clockPin, uint8_t ledsCount);
+        APA102(gpio_num_t mosiPin, gpio_num_t clockPin, size_t ledsCount);
         ~APA102();
         void flush();
         void setPixel(uint8_t index, PixelInfo pixel);
